server/requirements.txt
Flask
Flask-CORS
SQLAlchemy
python-dotenv
requests
python-magic # For file type detection